vihaco::binary

Trait BytecodeContext

pub trait BytecodeContext: Sized {
    // Required methods
    fn from_bytes(bytes: &[u8]) -> Result<Self>;
    fn section_name(&self, index: u32) -> Option<&str>;
}

The global context for a given program.

This should include all context needed for an entire section tree. Anything that should be shared across machines should be in a BytecodeContext.

Required Methods

fn from_bytes(bytes: &[u8]) -> Result<Self>

fn section_name(&self, index: u32) -> Option<&str>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.