vihaco::binary

Type Alias BinaryBytecodeFile

pub type BinaryBytecodeFile<C = ProgramContext> = BytecodeFile<Vec<u8>, C>;

Aliased Type

pub struct BinaryBytecodeFile<C = ProgramContext> { /* private fields */ }

Implementations

impl<C> BinaryBytecodeFile<C>
where C: BytecodeContext,

pub fn from_bytes(bytes: Vec<u8>) -> Result<BinaryBytecodeFile<C>>

The public entry point for the parsing of a bytecode file.

This will automatically split a multi-section file into each individual section node and send it to its corresponding composite loader marked with #[program].