vihaco::loader

Trait LoadBytecodeSection

pub trait LoadBytecodeSection<C> {
    // Required method
    fn load_bytecode_section<'bc>(
        &mut self,
        section: BytecodeSectionView<'bc, C>,
    ) -> Result<()>;
}

Allow a machine to load a bytecode section completely.

For composites generated by vihaco_derive::composite, this loads the composite’s own section through LoadOwnBytecodeSection and then forwards direct child sections to #[loadable] devices.

Required Methods

fn load_bytecode_section<'bc>( &mut self, section: BytecodeSectionView<'bc, C>, ) -> Result<()>