vihaco::machine

Trait ProgramCounter

pub trait ProgramCounter {
    type Instruction: Instruction;

    // Required methods
    fn pc(&self) -> u32;
    fn pc_mut(&mut self) -> &mut u32;
    fn get_instruction(&self, pc: u32) -> Result<&Self::Instruction>;

    // Provided methods
    fn peek_instruction(&self) -> Result<&Self::Instruction> { ... }
    fn next_instruction(&mut self) -> Result<&Self::Instruction> { ... }
}

Required Associated Types

type Instruction: Instruction

Required Methods

fn pc(&self) -> u32

Get the current program counter.

fn pc_mut(&mut self) -> &mut u32

Get a mutable reference to the program counter.

fn get_instruction(&self, pc: u32) -> Result<&Self::Instruction>

Get the instruction at the given program counter without advancing it.

Provided Methods

fn peek_instruction(&self) -> Result<&Self::Instruction>

Peek the current instruction without advancing the program counter.

fn next_instruction(&mut self) -> Result<&Self::Instruction>

Get the current instruction and advance the program counter by 1.