vihaco::value

Trait Value

pub trait Value:
    Sized
    + ByteWidth
    + FromBytes
    + WriteBytes {
    type Type: Type;

    // Required method
    fn type_of(&self) -> Self::Type;
}

A value for a vihaco component.

Values must be associated with some Type, connected by both Value::Type and Value::type_of.

Required Associated Types

type Type: Type

The type universe that this value set is contained within.

Required Methods

fn type_of(&self) -> Self::Type

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.